requirements/base.txt
# Bleeding edge Django
django==1.8.4

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.3


# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.6.0

# Models
django-model-utils==2.4

# Images
Pillow==3.1.0
six==1.10.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.24.1


# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.7

# Redis support
django-redis==4.3.0
redis>=2.10.0

celery==3.1.18

# Your custom requirements go here
Pygments==2.0.2
markdown
diff-match-patch

# Django REST Framework
djangorestframework==3.3.2
djangorestframework-recursive
drf-tracking==0.2.5

# Django tools
django-filter==0.11.0
django-mptt==0.8.0
django-avatar==2.2.0
django-annoying==0.8.5
django-messages-extends==0.4
django-notifications-hq==1.0
django-activity-stream==0.6.0
django-bootstrap-pagination==1.5.1
django-reversion==1.10.0
django-reversion-compare==0.5.6
django-licenses==0.2.5
django-grappelli==2.7.2
django-filebrowser==3.6.1

# Authentication
djoser==0.4.0
django-oauth-toolkit==0.10.0

# Comments
django-contrib-comments==1.6.2
django-threadedcomments==1.0.1
django-fluent-comments==1.1

# MongoDB
pymongo
djangotoolbox
mongoengine

# Google Plus
google-api-python-client==1.3.1
httplib2==0.9.1
